+++ plugins/woocommerce-admin/client/header/activity-panel/panels/inbox.js
@@ -18,6 +18,7 @@ import { EmptyContent, Section } from '@woocommerce/components';
 import sanitizeHTML from 'lib/sanitize-html';
 import { QUERY_DEFAULTS } from 'wc-api/constants';
 import withSelect from 'wc-api/with-select';
+import classnames from 'classnames';
 
 class InboxPanel extends Component {
 	constructor( props ) {
@@ -49,18 +50,23 @@ class InboxPanel extends Component {
 	}
 
 	renderNotes() {
-		const { lastRead, notes } = this.props;
+		const { lastRead, notes, triggerNoteAction } = this.props;
 
 		if ( 0 === Object.keys( notes ).length ) {
 			return this.renderEmptyCard();
 		}
 
-		const getButtonsFromActions = actions => {
-			if ( ! actions ) {
+		const getButtonsFromActions = note => {
+			if ( ! note.actions ) {
 				return [];
 			}
-			return actions.map( action => (
-				<Button isDefault href={ action.url || undefined }>
+			return note.actions.map( action => (
+				<Button
+					isDefault
+					isPrimary={ action.primary }
+					href={ action.url || undefined }
+					onClick={ () => triggerNoteAction( note.id, action.id ) }
+				>
 					{ action.label }
 				</Button>
 			) );
@@ -71,7 +77,9 @@ class InboxPanel extends Component {
 		return notesArray.map( note => (
 			<ActivityCard
 				key={ note.id }
-				className="woocommerce-inbox-activity-card"
+				className={ classnames( 'woocommerce-inbox-activity-card', {
+					actioned: 'unactioned' !== note.status,
+				} ) }
 				title={ note.title }
 				date={ note.date_created_gmt }
 				icon={ <Gridicon icon={ note.icon } size={ 48 } /> }
@@ -80,7 +88,7 @@ class InboxPanel extends Component {
 					! note.date_created_gmt ||
 					new Date( note.date_created_gmt + 'Z' ).getTime() > lastRead
 				}
-				actions={ getButtonsFromActions( note.actions ) }
+				actions={ getButtonsFromActions( note ) }
 			>
 				<span dangerouslySetInnerHTML={ sanitizeHTML( note.content ) } />
 			</ActivityCard>
@@ -145,6 +153,7 @@ export default compose(
 			type: 'info,warning',
 			orderby: 'date',
 			order: 'desc',
+			status: 'unactioned',
 		};
 
 		const notes = getNotes( inboxQuery );
@@ -154,10 +163,11 @@ export default compose(
 		return { notes, isError, isRequesting, lastRead: userData.activity_panel_inbox_last_read };
 	} ),
 	withDispatch( dispatch => {
-		const { updateCurrentUserData } = dispatch( 'wc-api' );
+		const { updateCurrentUserData, triggerNoteAction } = dispatch( 'wc-api' );
 
 		return {
 			updateCurrentUserData,
+			triggerNoteAction,
 		};
 	} )
 )( InboxPanel );