+++ plugins/woocommerce-admin/client/wc-api/notes/mutations.js
@@ -5,6 +5,12 @@ const updateNote = operations => ( noteId, noteFields ) => {
 	operations.update( [ resourceKey ], { [ resourceKey ]: { noteId, ...noteFields } } );
 };
 
+const triggerNoteAction = operations => ( noteId, actionId ) => {
+	const resourceKey = 'note-action';
+	operations.update( [ resourceKey ], { [ resourceKey ]: { noteId, actionId } } );
+};
+
 export default {
 	updateNote,
+	triggerNoteAction,
 };