+++ plugins/woocommerce-admin/includes/class-wc-admin-install.php
@@ -94,6 +94,7 @@ class WC_Admin_Install {
 			order_id bigint(20) unsigned NOT NULL,
 			parent_id bigint(20) unsigned DEFAULT 0 NOT NULL,
 			date_created datetime DEFAULT '0000-00-00 00:00:00' NOT NULL,
+			date_created_gmt datetime DEFAULT '0000-00-00 00:00:00' NOT NULL,
 			num_items_sold int(11) DEFAULT 0 NOT NULL,
 			gross_total double DEFAULT 0 NOT NULL,
 			tax_total double DEFAULT 0 NOT NULL,
@@ -271,6 +272,19 @@ class WC_Admin_Install {
 	protected static function create_notes() {
 		WC_Admin_Notes_Historical_Data::add_note();
 		WC_Admin_Notes_Welcome_Message::add_welcome_note();
+	}
+
+	/**
+	 * Delete all data from tables.
+	 */
+	public static function delete_table_data() {
+		global $wpdb;
+
+		$tables = self::get_tables();
+
+		foreach ( $tables as $table ) {
+			$wpdb->query( "TRUNCATE TABLE {$table}" ); // WPCS: unprepared SQL ok.
+		}
 	}
 }