+++ plugins/woocommerce-admin/tests/api/reports-import.php
@@ -330,8 +330,9 @@ class WC_Tests_API_Reports_Import extends WC_REST_Unit_Test_Case {
 		$report     = $response->get_data();
 		$user_query = new WP_User_Query(
 			array(
-				'fields' => 'ID',
-				'number' => 1,
+				'fields'   => 'ID',
+				'number'   => 1,
+				'role__in' => array( 'customer' ),
 			)
 		);
 
@@ -363,7 +364,7 @@ class WC_Tests_API_Reports_Import extends WC_REST_Unit_Test_Case {
 		$this->assertEquals( 200, $response->get_status() );
 		$this->assertEquals( true, $report['is_importing'] );
 		$this->assertEquals( 0, $report['customers_count'] );
-		$this->assertEquals( 3, $report['customers_total'] );
+		$this->assertEquals( 1, $report['customers_total'] );
 		$this->assertEquals( 0, $report['orders_count'] );
 		$this->assertEquals( 4, $report['orders_total'] );
 
@@ -380,7 +381,7 @@ class WC_Tests_API_Reports_Import extends WC_REST_Unit_Test_Case {
 		$this->assertEquals( 200, $response->get_status() );
 		$this->assertEquals( false, $report['is_importing'] );
 		$this->assertEquals( 1, $report['customers_count'] );
-		$this->assertEquals( 3, $report['customers_total'] );
+		$this->assertEquals( 1, $report['customers_total'] );
 		$this->assertEquals( 4, $report['orders_count'] );
 		$this->assertEquals( 4, $report['orders_total'] );
 
@@ -390,8 +391,7 @@ class WC_Tests_API_Reports_Import extends WC_REST_Unit_Test_Case {
 		$response = $this->server->dispatch( $request );
 		$report   = $response->get_data();
 
-		// @todo The following line should be uncommented when https://github.com/woocommerce/woocommerce-admin/issues/2195 is resolved.
-		// $this->assertEquals( 0, $report['customers'] );
+		$this->assertEquals( 0, $report['customers'] );
 		$this->assertEquals( 0, $report['orders'] );
 	}
 }